Use float for angles instead + syntax fixes

+++ src/encoderTask.cpp
@@ -14,7 +14,7 @@ std::array<AS5600, N_ENCODERS> encoderArray;
 std::array<float, N_ENCODERS> encoderPositionArray;
 
 // === EXTERNALS === //
-std::array<int, N_ENCODERS> currentAngleArray;
+std::array<float, N_ENCODERS> currentAngleArray;
 
 // === INTERRUPT === //
